+++ Jellyfin.Plugin.JellyfinHelper/Plugin.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
+using System.Reflection;
 using Jellyfin.Plugin.JellyfinHelper.Configuration;
 using MediaBrowser.Common.Configuration;
 using MediaBrowser.Common.Plugins;
@@ -38,6 +40,16 @@ public class Plugin : BasePlugin<PluginConfiguration>, IHasWebPages
     /// </summary>
     public static Plugin? Instance { get; private set; }
 
+    /// <summary>
+    /// Gets the plugin thumb image stream from the embedded resource.
+    /// </summary>
+    /// <returns>The image stream, or null if the resource is not found.</returns>
+    public Stream? GetThumbImage()
+    {
+        return Assembly.GetExecutingAssembly().GetManifestResourceStream(
+            GetType().Namespace + ".logo.png");
+    }
+
     /// <inheritdoc />
     public IEnumerable<PluginPageInfo> GetPages()
     {
@@ -48,6 +60,7 @@ public class Plugin : BasePlugin<PluginConfiguration>, IHasWebPages
                 Name = Name,
                 DisplayName = "Jellyfin Helper",
                 EnableInMainMenu = true,
+                MenuIcon = "handyman",
                 EmbeddedResourcePath = GetType().Namespace + ".Configuration.configPage.html"
             }
         ];